requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
openai==1.12.0
python-dotenv==1.0.1
jinja2==3.1.3
aiofiles==23.2.1
pymupdf==1.24.0
pillow==10.2.0
reportlab==4.1.0
pdf2docx==0.5.8
sqlalchemy>=2.0.46
psycopg2>=2.9.9
bcrypt>=4.0.0
itsdangerous>=2.1.2
authlib>=1.2.0
httpx>=0.24.0
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
chromadb>=0.3.21
sentence-transformers>=2.2.2